hospital_web/requirement.txt
psycopg2
django
django-widget-tweaks
djangorestframework
django-cors-headers
Pillow
requests